raft: Split out fundamental and replicate from types
