+++ src/v/raft/fundamental.h
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2024 Redpanda Data, Inc.
+ *
+ * Use of this software is governed by the Business Source License
+ * included in the file licenses/BSL.md
+ *
+ * As of the Change Date specified in that file, in accordance with
+ * the Business Source License, use of this software will be governed
+ * by the Apache License, Version 2.0
+ */
+
+#pragma once
+
+#include "base/seastarx.h"
+#include "utils/named_type.h"
+
+#include <seastar/core/lowres_clock.hh>
+
+namespace raft {
+
+using clock_type = ss::lowres_clock;
+using duration_type = typename clock_type::duration;
+using timer_type = ss::timer<clock_type>;
+static constexpr clock_type::time_point no_timeout
+  = clock_type::time_point::max();
+
+using group_id = named_type<int64_t, struct raft_group_id_type>;
+
+enum class reply_result : uint8_t {
+    success,
+    failure,
+    group_unavailable,
+    timeout
+};
+
+} // namespace raft
